% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentAssignFunc.R
\name{assignStudents}
\alias{assignStudents}
\title{Assigns students to schools}
\usage{
assignStudents(kg, minA = 20, minB = 10, minC = 20, minH = 5,
  maxA = 40, maxB = 40, maxC = 60, maxH = 1000, penalty = 0)
}
\arguments{
\item{kg}{a data frame containing the nx4 distance matrix}

\item{minA}{minimum number of kindergarten students to assign to Ames}

\item{minB}{minimum number of kindergarten students to assign to Blythe}

\item{minC}{minimum number of kindergarten students to assign to Central}

\item{minH}{minimum number of kindergarten students to assign to Hollywood}

\item{maxA}{maximum number of kindergarten students to assign to Ames}

\item{maxB}{maximum number of kindergarten students to assign to Blythe}

\item{maxC}{maximum number of kindergarten students to assign to Central}

\item{maxH}{maximum number of kindergarten students to assign to Hollywood}

\item{penalty}{adds a term that is quadtaric in distance squared to the walking distance matrix. makes solution more compact}
}
\value{
assignment of students to schools which minimizes total walking distance subject to space constraints.
}
\description{
\code{assignStudents} assigns students to schools to minimize total walking distance subject to constraints on class sizes and
the number of sections per school. The user specifies the minimum and maximum number of students that can be assigned to each school.
If \code{penalty} is not zero, the function that is minimized is the sum of all walking distances plus \code{penalty} times the sum of all walking distances squared
}
