% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistMatrix.R
\name{getDistanceMatrix}
\alias{getDistanceMatrix}
\title{Uses Google to get walking times for each student to each school. Also geocodes addresses}
\usage{
getDistanceMatrix(enrollDF)
}
\arguments{
\item{enrollDF}{a data frame containing the nx4 distance matrix}
}
\description{
\code{getDistanceMatrix} Calls the Google API to get walkig times in minutes for each student to each school and also geocodes
each address. Requires a Google key. Google allows 2500 free queries per day. The function creates an Excel spreadsheet
"enrollDFwDist-(time stamp).xlsx containing all the info in the input object (enrollDF) plus six new columns: the
distance to each school (in minutes), the latitude and the longitude of each student address
}
