% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d96plotfunctions.R
\name{plotAssign}
\alias{plotAssign}
\title{Plots student assignments}
\usage{
plotAssign(sa, what = "AlgoAssign", ptitle = "", pcaption = "",
  dsize = 1, print_map = T, djitter = 3e-04)
}
\arguments{
\item{sa}{a student assignment dataframe}

\item{what}{kind of assignment to be plotted. Options are "AlgoAssign", "School" (actual assignment), "Final" (same as Algo except for Siblings)}

\item{ptitle}{title for plot}

\item{pcaption}{caption for plot}

\item{dsize}{size of points}

\item{print_map}{whether to print the map or return a ggmap object}

\item{djitter}{amount of jitter}
}
\value{
a plot showing how students are assigned
}
\description{
\code{plotAssign} shows  on a map how students are assigned to schools. There are a number of options for what can be plotted,
inlcuding how the algorithm assigns students, the actual assignment of students, etc.
}
