% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{riversideMapBW}
\alias{riversideMapBW}
\title{Another map of D96 area}
\format{An object of class \code{gg} (inherits from \code{ggplot}) of length 9.}
\usage{
riversideMapBW
}
\description{
Black and white Google Map of D96 area
}
\keyword{datasets}
