% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tschoolGeo}
\alias{tschoolGeo}
\title{Elementary school locations}
\format{A data frame with 4 rows and 4 variables:
\describe{
  \item{address}{address}
  \item{lon}{longitude}
  \item{lat}{latitude}
  \item{school}{school name}
}}
\usage{
tschoolGeo
}
\description{
Address, lat, and lon of four elementary schools.
}
\keyword{datasets}
